+++ rich/theme.py
@@ -1,5 +1,5 @@
 import configparser
-from typing import IO, Mapping
+from typing import Dict, List, IO, Mapping, Optional
 
 from .default_styles import DEFAULT_STYLES
 from .style import Style, StyleType
@@ -9,9 +9,11 @@ class Theme:
     """A container for style information, used by :class:`~rich.console.Console`.
 
     Args:
-        styles (Dict[str, Style], optional): A mapping of style names on to styles. Defaults to None for empty styles.
+        styles (Dict[str, Style], optional): A mapping of style names on to styles. Defaults to None for a theme with no styles.
         inherit (bool, optional): Inherit default styles. Defaults to True.
     """
+
+    styles: Dict[str, Style]
 
     def __init__(self, styles: Mapping[str, StyleType] = None, inherit: bool = True):
         self.styles = DEFAULT_STYLES.copy() if inherit else {}
@@ -64,6 +66,43 @@ class Theme:
         """
         with open(path, "rt") as config_file:
             return cls.from_file(config_file, source=path, inherit=inherit)
+
+
+class ThemeStackError(Exception):
+    """Base exception for errors related to the theme stack."""
+
+
+class ThemeStack:
+    """A stack of themes.
+
+    Args:
+        theme (Theme): A theme instance
+    """
+
+    def __init__(self, theme: Theme) -> None:
+        self._entries: List[Dict[str, Style]] = [theme.styles]
+        self.get = self._entries[-1].get
+
+    def push_theme(self, theme: Theme, inherit: bool = True) -> None:
+        """Push a theme on the top of the stack.
+
+        Args:
+            theme (Theme): A Theme instance.
+            inherit (boolean, optional): Inherit styles from current top of stack.
+        """
+        styles: Dict[str, Style]
+        styles = (
+            {**self._entries[-1], **theme.styles} if inherit else theme.styles.copy()
+        )
+        self._entries.append(styles)
+        self.get = self._entries[-1].get
+
+    def pop_theme(self) -> None:
+        """Pop (and discard) the top-most theme."""
+        if len(self._entries) == 1:
+            raise ThemeStackError("Unable to pop base theme")
+        self._entries.pop()
+        self.get = self._entries[-1].get
 
 
 if __name__ == "__main__":  # pragma: no cover