+++ rich/traceback.py
@@ -3,13 +3,26 @@ from __future__ import absolute_import
 import platform
 import sys
 from dataclasses import dataclass, field
-from traceback import extract_tb
+from traceback import walk_tb
 from types import TracebackType
 from typing import Callable, Dict, List, Optional, Type
 
-from pygments.token import Token, String, Name, Number
+from pygments.lexers import guess_lexer_for_filename
+from pygments.token import (
+    Comment,
+    Generic,
+    Keyword,
+    Name,
+    Number,
+    Operator,
+    String,
+    Token,
+    Text as TextToken,
+)
 
+from . import pretty
 from ._loop import loop_first, loop_last
+from .columns import Columns
 from .console import (
     Console,
     ConsoleOptions,
@@ -19,11 +32,12 @@ from .console import (
 )
 from .constrain import Constrain
 from .highlighter import RegexHighlighter, ReprHighlighter
-from .padding import Padding
 from .panel import Panel
-from .syntax import Syntax
+from .scope import render_scope
 from .style import Style
+from .syntax import Syntax
 from .text import Text
+from .theme import Theme
 
 WINDOWS = platform.system() == "Windows"
 
@@ -35,6 +49,7 @@ def install(
     extra_lines: int = 3,
     theme: Optional[str] = None,
     word_wrap: bool = False,
+    show_locals: bool = False,
 ) -> Callable:
     """Install a rich traceback handler.
 
@@ -47,7 +62,8 @@ def install(
         extra_lines (int, optional): Extra lines of code. Defaults to 3.
         theme (Optional[str], optional): Pygments theme to use in traceback. Defaults to ``None`` which will pick
             a theme appropriate for the platform.
-        word_wrap(bool, optional): Enable word wrapping of long lines. Defaults to False.
+        word_wrap (bool, optional): Enable word wrapping of long lines. Defaults to False.
+        show_locals (bool, optional): Enable display of local variables. Defaults to False.
 
     Returns:
         Callable: The previous exception handler that was replaced.
@@ -69,6 +85,7 @@ def install(
                 extra_lines=extra_lines,
                 theme=theme,
                 word_wrap=word_wrap,
+                show_locals=show_locals,
             )
         )
 
@@ -83,6 +100,7 @@ class Frame:
     lineno: int
     name: str
     line: str = ""
+    locals: Optional[Dict[str, pretty.Node]] = None
 
 
 @dataclass
@@ -121,6 +139,7 @@ class Traceback:
         extra_lines (int, optional): Additional lines of code to render. Defaults to 3.
         theme (str, optional): Override pygments theme used in traceback.
         word_wrap (bool, optional): Enable word wrapping of long lines. Defaults to False.
+        show_locals (bool, optional): Enable display of local variables. Defaults to False.
     """
 
     def __init__(
@@ -130,6 +149,7 @@ class Traceback:
         extra_lines: int = 3,
         theme: Optional[str] = None,
         word_wrap: bool = False,
+        show_locals: bool = False,
     ):
         if trace is None:
             exc_type, exc_value, traceback = sys.exc_info()
@@ -137,12 +157,15 @@ class Traceback:
                 raise ValueError(
                     "Value for 'trace' required if not called in except: block"
                 )
-            trace = self.extract(exc_type, exc_value, traceback)
+            trace = self.extract(
+                exc_type, exc_value, traceback, show_locals=show_locals
+            )
         self.trace = trace
         self.width = width
         self.extra_lines = extra_lines
         self.theme = Syntax.get_theme(theme or "ansi_dark")
         self.word_wrap = word_wrap
+        self.show_locals = show_locals
 
     @classmethod
     def from_exception(
@@ -154,6 +177,7 @@ class Traceback:
         extra_lines: int = 3,
         theme: Optional[str] = None,
         word_wrap: bool = False,
+        show_locals: bool = False,
     ) -> "Traceback":
         """Create a traceback from exception info
 
@@ -165,6 +189,7 @@ class Traceback:
             extra_lines (int, optional): Additional lines of code to render. Defaults to 3.
             theme (str, optional): Override pygments theme used in traceback.
             word_wrap (bool, optional): Enable word wrapping of long lines. Defaults to False.
+            show_locals (bool, optional): Enable display of local variables. Defaults to False.
 
         Returns:
             Traceback: A Traceback instance that may be printed.
@@ -176,6 +201,7 @@ class Traceback:
             extra_lines=extra_lines,
             theme=theme,
             word_wrap=word_wrap,
+            show_locals=show_locals,
         )
 
     @classmethod
@@ -184,6 +210,7 @@ class Traceback:
         exc_type: Type[BaseException],
         exc_value: BaseException,
         traceback: Optional[TracebackType],
+        show_locals: bool = False,
     ) -> Trace:
         """Extrace traceback information.
 
@@ -191,10 +218,12 @@ class Traceback:
             exc_type (Type[BaseException]): Exception type.
             exc_value (BaseException): Exception value.
             traceback (TracebackType): Python Traceback object.
+            show_locals (bool, optional): Enable display of local variables. Defaults to False.
 
         Returns:
             Trace: A Trace instance which you can use to construct a `Traceback`.
         """
+
         stacks: List[Stack] = []
         while True:
             stack = Stack(exc_type=str(exc_type.__name__), exc_value=str(exc_value))
@@ -211,11 +240,17 @@ class Traceback:
             stacks.append(stack)
             append = stack.frames.append
 
-            for frame_summary in extract_tb(traceback):
+            for frame_summary, line_no in walk_tb(traceback):
                 frame = Frame(
-                    filename=frame_summary.filename,
-                    lineno=frame_summary.lineno,
-                    name=frame_summary.name,
+                    filename=frame_summary.f_code.co_filename or "?",
+                    lineno=line_no,
+                    name=frame_summary.f_code.co_name,
+                    locals={
+                        key: pretty.traverse(value)
+                        for key, value in frame_summary.f_locals.items()
+                    }
+                    if show_locals
+                    else None,
                 )
                 append(frame)
 
@@ -237,18 +272,33 @@ class Traceback:
     ) -> RenderResult:
         theme = self.theme
         background_style = theme.get_background_style()
+        token_style = theme.get_style_for_token
 
-        styles = {
-            "text": theme.get_style_for_token(Token),
-            "string": theme.get_style_for_token(String),
-            "function": theme.get_style_for_token(Name.Function),
-            "number": theme.get_style_for_token(Number),
-        }
+        traceback_theme = Theme(
+            {
+                "pretty": token_style(TextToken),
+                "pygments.text": token_style(Token),
+                "pygments.string": token_style(String),
+                "pygments.function": token_style(Name.Function),
+                "pygments.number": token_style(Number),
+                "repr.str": token_style(String),
+                "repr.brace": token_style(TextToken) + Style(bold=True),
+                "repr.number": token_style(Number),
+                "repr.bool_true": token_style(Keyword.Constant),
+                "repr.bool_false": token_style(Keyword.Constant),
+                "repr.none": token_style(Keyword.Constant),
+                "scope.border": token_style(String.Delimiter),
+                "scope.equals": token_style(Operator),
+                "scope.key": token_style(Name),
+                "scope.key.special": token_style(Name.Constant) + Style(dim=True),
+            }
+        )
+
         highlighter = ReprHighlighter()
         for last, stack in loop_last(reversed(self.trace.stacks)):
             if stack.frames:
                 stack_renderable: ConsoleRenderable = Panel(
-                    self._render_stack(stack, styles),
+                    self._render_stack(stack),
                     title="[traceback.title]Traceback [dim](most recent call last)",
                     style=background_style,
                     border_style="traceback.border.syntax_error",
@@ -256,18 +306,20 @@ class Traceback:
                     padding=(0, 1),
                 )
                 stack_renderable = Constrain(stack_renderable, self.width)
-                yield stack_renderable
+                with console.use_theme(traceback_theme):
+                    yield stack_renderable
             if stack.syntax_error is not None:
-                yield Constrain(
-                    Panel(
-                        self._render_syntax_error(stack.syntax_error, styles),
-                        style=background_style,
-                        border_style="traceback.border",
-                        expand=False,
-                        padding=(0, 1),
-                    ),
-                    self.width,
-                )
+                with console.use_theme(traceback_theme):
+                    yield Constrain(
+                        Panel(
+                            self._render_syntax_error(stack.syntax_error),
+                            style=background_style,
+                            border_style="traceback.border",
+                            expand=False,
+                            padding=(0, 1),
+                        ),
+                        self.width,
+                    )
                 yield Text.assemble(
                     (f"{stack.exc_type}: ", "traceback.exc_type"),
                     highlighter(stack.syntax_error.msg),
@@ -284,17 +336,15 @@ class Traceback:
                 )
 
     @render_group()
-    def _render_syntax_error(
-        self, syntax_error: _SyntaxError, styles: Dict[str, Style]
-    ) -> RenderResult:
+    def _render_syntax_error(self, syntax_error: _SyntaxError) -> RenderResult:
         highlighter = ReprHighlighter()
         path_highlighter = PathHighlighter()
         if syntax_error.filename != "<stdin>":
             text = Text.assemble(
-                (f" {syntax_error.filename}", styles["string"]),
-                (":", styles["text"]),
-                (str(syntax_error.lineno), styles["number"]),
-                style=styles["text"],
+                (f" {syntax_error.filename}", "pygments.string"),
+                (":", "pgments.text"),
+                (str(syntax_error.lineno), "pygments.number"),
+                style="pygments.text",
             )
             yield path_highlighter(text)
         syntax_error_text = highlighter(syntax_error.line.rstrip())
@@ -303,22 +353,40 @@ class Traceback:
         syntax_error_text.stylize("bold underline", offset, offset + 1)
         syntax_error_text += Text.from_markup(
             "\n" + " " * offset + "[traceback.offset]▲[/]",
-            style=styles["text"],
+            style="pygments.text",
         )
         yield syntax_error_text
 
     @render_group()
-    def _render_stack(self, stack: Stack, styles: Dict[str, Style]) -> RenderResult:
+    def _render_stack(self, stack: Stack) -> RenderResult:
         path_highlighter = PathHighlighter()
         theme = self.theme
+        code_cache: Dict[str, str] = {}
+
+        def read_code(filename: str) -> str:
+            """Read files, and cache results on filename.
+
+            Args:
+                filename (str): Filename to read
+
+            Returns:
+                str: Contents of file
+            """
+            code = code_cache.get(filename)
+            if code is None:
+                with open(filename, "rt") as code_file:
+                    code = code_file.read()
+                code_cache[filename] = code
+            return code
+
         for first, frame in loop_first(stack.frames):
             text = Text.assemble(
-                path_highlighter(Text(frame.filename, style=styles["string"])),
-                (":", styles["text"]),
-                (str(frame.lineno), styles["number"]),
+                path_highlighter(Text(frame.filename, style="pygments.string")),
+                (":", "pygments.text"),
+                (str(frame.lineno), "pygments.number"),
                 " in ",
-                (frame.name, styles["function"]),
-                style=styles["text"],
+                (frame.name, "pygments.function"),
+                style="pygments.text",
             )
             if not frame.filename.startswith("<") and not first:
                 yield ""
@@ -326,8 +394,12 @@ class Traceback:
             if frame.filename.startswith("<"):
                 continue
             try:
-                syntax = Syntax.from_path(
-                    frame.filename,
+                code = read_code(frame.filename)
+                lexer = guess_lexer_for_filename(frame.filename, code)
+                lexer_name = lexer.name
+                syntax = Syntax(
+                    code,
+                    lexer_name,
                     theme=theme,
                     line_numbers=True,
                     line_range=(
@@ -336,12 +408,23 @@ class Traceback:
                     ),
                     highlight_lines={frame.lineno},
                     word_wrap=self.word_wrap,
+                    code_width=88,
                 )
                 yield ""
             except Exception:
                 pass
             else:
-                yield syntax
+                yield (
+                    Columns(
+                        [
+                            syntax,
+                            render_scope(frame.locals, title="locals"),
+                        ],
+                        padding=1,
+                    )
+                    if frame.locals
+                    else syntax
+                )
 
 
 if __name__ == "__main__":  # pragma: no cover
@@ -352,17 +435,29 @@ if __name__ == "__main__":  # pragma: no cover
     import sys
 
     def bar(a):  # 这是对亚洲语言支持的测试。面对模棱两可的想法，拒绝猜测的诱惑
-        print(1 / a)
+        one = 1
+        print(one / a)
 
     def foo(a):
+        zed = {
+            "characters": {
+                "Paul Atriedies",
+                "Vladimir Harkonnen",
+                "Thufir Haway",
+                "Duncan Idaho",
+            },
+            "atomic_types": (None, False, True),
+        }
         bar(a)
 
-    try:
+    def error():
+
         try:
-            foo(0)
+            try:
+                foo(0)
+            except:
+                slfkjsldkfj  # type: ignore
         except:
-            slfkjsldkfj  # type: ignore
-    except:
-        tb = Traceback()
-        # print(fooads)
-        console.print(tb)
+            console.print_exception(show_locals=True)
+
+    error()