+++ tests/test_logging.py
@@ -25,19 +25,6 @@ skip_win = pytest.mark.skipif(
 )
 
 
-def make_log():
-    log.debug("foo")
-    render = handler.console.file.getvalue()
-    return render
-
-
-def test_log():
-    render = make_log()
-    print(repr(render))
-    expected = "\x1b[2;36m[DATE]\x1b[0m\x1b[2;36m \x1b[0m\x1b[32mDEBUG\x1b[0m    foo                                           \x1b[2mtest_logging.py\x1b[0m\x1b[2m:29\x1b[0m\n"
-    assert render == expected
-
-
 @skip_win
 def test_exception():
     console = Console(