+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.0.0] - 2020-10-03
+
+### Added
+
+- Added Console.bell method
+- Added Set to types that Console.print will automatically pretty print
+- Added show_locals to Traceback
+- Added theme stack mechanism, see Console.push_theme and Console.pop_theme
+
+### Changed
+
+- Changed Style.empty to Style.null to better reflect what it does
+- Optimized combining styles involving a null style
+- Change error messages in Style.parse to read better
+
+### Fixed
+
+- Fixed Table.\_\_rich_measure\_\_
+- Fixed incorrect calculation of fixed width columns
+
 ## [7.1.0] - 2020-09-26
 
 ### Added